[CXF-7571] Migrate context object creation into the bean.
