+++ integration/cdi/src/main/java/org/apache/cxf/cdi/AbstractCXFBean.java
@@ -25,9 +25,11 @@ import java.util.HashSet;
 import java.util.Set;
 
 import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.spi.CreationalContext;
 import javax.enterprise.inject.Any;
 import javax.enterprise.inject.Default;
 import javax.enterprise.inject.spi.Bean;
+import javax.enterprise.inject.spi.InjectionPoint;
 import javax.enterprise.util.AnnotationLiteral;
 
 abstract class AbstractCXFBean<T> implements Bean<T> {
@@ -63,6 +65,17 @@ abstract class AbstractCXFBean<T> implements Bean<T> {
     @Override
     public boolean isNullable() {
         return false;
+    }
+
+    @Override
+    public Set<InjectionPoint> getInjectionPoints() {
+        return Collections.emptySet();
+    }
+
+
+    @Override
+    public void destroy(T instance, CreationalContext<T> creationalContext) {
+        creationalContext.release();
     }
 
     @Override
